+++ R/odds.ratio.glm.R
@@ -2,6 +2,7 @@
 #' @title Calculate Odds Ratio of Generalized Linear (Mixed) Models
 #' 
 #' @importFrom stats coefficients
+#' @importFrom stats confint
 #' 
 #' @description This function calculates odds ratio(s) for specific 
 #'     increment steps of GLMs. 
@@ -9,11 +10,19 @@
 #' @param data The data used for model fitting.
 #' @param model A fitted GLM(M).
 #' @param incr List. Increment values of each predictor.
+#' @param CI numeric. Which confident interval to calculate. 
+#' Must be between 0 and 1. Default to 0.95
 #' 
-#' @return A data frame with three columns:
+#' @return A data frame with five columns:
 #' \item{predictor}{Predictor name(s)}
 #' \item{oddsratio}{Calculated odds ratio(s)}
+#' \item{CI.low}{Lower confident interval of odds ratio}
+#' \item{CI.high}{Higher confident interval of odds ratio}
 #' \item{increment}{Increment of the predictor(s)}
+#' 
+#' @details \code{CI.low} and {CI.high} are only calculted for GLM models because 
+#' \code{\link[MASS]{glmmPQL}} does not return confident intervals due to its penalizing
+#' behaviour. 
 #' 
 #' @examples 
 #' ## Example with stats::glm()
@@ -24,6 +33,10 @@
 #' 
 #' # Calculate OR for specific increment step of continuous variable
 #' calc.oddsratio.glm(data = dat, model = fit.glm, incr = list(gre = 380, gpa = 5))
+#' 
+#' # Calculate OR and change the confidence interval level
+#' calc.oddsratio.glm(data = dat, model = fit.glm, 
+#'                    incr = list(gre = 380, gpa = 5), CI = .70)
 #' 
 #' ## Example with MASS:glmmPQL()
 #' # load data
@@ -41,7 +54,7 @@
 #' @seealso \code{\link[oddsratio]{calc.oddsratio.gam}}
 #' 
 #' @export
-calc.oddsratio.glm <- function(data, model, incr) {
+calc.oddsratio.glm <- function(data, model, incr, CI = 0.95) {
   
   if (class(model)[1] == "glm") {
     # get pred names and coefficients without intercept
@@ -53,30 +66,65 @@ calc.oddsratio.glm <- function(data, model, incr) {
     # get pred names and coefficients without intercept
     preds <- names(model$coefficients$fixed)[2:length(model$coefficients$fixed)]
     coef <- model$coefficients$fixed[2:length(model$coefficients$fixed)]
+    cat("Warning: No confident interval calculation possible 
+        for 'glmmPQL' models\n\n")
   }
   
   increments <- list()
   odds.ratios <- list()
+  CI.low <- list()
+  CI.high <- list()
   for (i in preds) {
+    
+    # CI calculation
+    if (class(model)[1] == "glm") {
+      CI.list <- as.data.frame(suppressMessages(confint(model, level = CI))) [-1, ] 
+    }
+    
     # check if predictor is numeric or integer
     if (is.numeric(data[[i]]) | is.integer(data[[i]])) {
       odds.ratios[[i]] <- round(exp(as.numeric(coef[[i]]) * as.numeric(incr[[i]])), 3)
+      if (!class(model)[1] == "glmmPQL") {
+        CI.low[[i]] <- round(exp(CI.list[i, 1] * as.numeric(incr[[i]])), 3)
+        CI.high[[i]] <- round(exp(CI.list[i, 2] * as.numeric(incr[[i]])), 3)
+      }
       increments[[i]] <- as.numeric(incr[[i]])
       or <- odds.ratios[[i]]
     }
     # if pred is factor -> perform direct conversion to odds ratio
     else {
       odds.ratios[[i]] <- round(exp(as.numeric(coef[[i]])), 3)
+      
+      if (!class(model)[1] == "glmmPQL") {
+        CI.low[[i]] <- round(exp(CI.list[i, 1]), 3)
+        CI.high[[i]] <- round(exp(CI.list[i, 2]), 3)
+      }
+      
       increments[[i]] <- "Indicator variable"
       or <- odds.ratios[[i]]
     }
   }
   
+  # set CIs NA if model is of type glmmPQL
+  if (class(model)[1] == "glmmPQL") {
+    CI.low <- c(rep(NA, length(preds)))
+    CI.high <- c(rep(NA, length(preds)))
+  }
+  
   # create data frame to return
   result <- data.frame(predictor = as.character(names(odds.ratios)),
                        oddsratio = unlist(odds.ratios, use.names = FALSE),
+                       CI.low = unlist(CI.low, use.names = FALSE),
+                       CI.high = unlist(CI.high, use.names = FALSE),
                        increment = as.character(unlist(increments, 
                                                        use.names = FALSE)),
                        stringsAsFactors = FALSE)
+  
+  # set CI column names
+  if (class(model)[1] == "glm") {
+    colnames(result)[3] <- paste0("CI.low (", names(CI.list) [1], ")")
+    colnames(result)[4] <- paste0("CI.high (", names(CI.list) [2], ")")
+  }
+  
   return(result)
 }