+++ NEWS.md
@@ -1,3 +1,9 @@
+# oddsratio 0.3 
+
+* Add confident interval calculation (`calc.oddsratio.glm`, `calc.oddsratio.gam`)
+* For GLM models CI level can be specified manually
+* Print CI warning if model is of type `glmmPQL`
+
 # oddsratio 0.2 (Oct 12 2016)
 
 * Remove param `quietly`